requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
duckduckgo_search==8.0.3
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.5.1
gradio==5.33.2
gradio_client==1.10.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
litellm==1.72.4
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
numpy==2.3.0
ollama==0.5.1
openai==1.86.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==11.1.0
primp==0.15.0
propcache==0.3.2
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
ruff==0.11.13
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smolagents==1.13.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
